feat(a11y): improve accessibility of headings popup
Headings popup now supports keyboard interactivity (via tab key)
and has menu/menuitem roles.

Note that changes are required beyond the scope of this commit for full
accessibility, such as an aria-haspopup/aria-controls on the menu
opener, but these require ID values, which appear to not be supported at
this time.