feat(a11y): add "toolbarStyle" to allow wrapping in smaller resos
The More/"..." button shown in smaller viewports violates the ARIA Menu
or Menu Bar Design Pattern. This includes the potential for launching
additional submenus, e.g., if the Header popup is launched from within
"...", which needs its own set of roles/states/keyboard interactivity
independent from its placement in the main toolbar. Instead of handling
all possible roles/states/keyboard interactivity, this commit adds an option,
"toolbarStyle", with an optional parameter of 'wrap', which allows
for visually wrapping the toolbar elements, such that the More button is
not shown, and thus the remaining toolbar elements only need to adhere
to the ARIA Toolbar Design Pattern.